refactor(Bot): rename translation unit
