+++ Classes/Bot.hpp
